onetomany mapping functionality added

+++ src/main/java/com/reallabs/eams/controller/LoanController.java
@@ -1,9 +1,12 @@
 package com.reallabs.eams.controller;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,39 +15,79 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.reallabs.eams.model.Account;
 import com.reallabs.eams.model.Loan;
+import com.reallabs.eams.service.AccountService;
 import com.reallabs.eams.service.LoanService;
 
 @RestController
 public class LoanController {
-	
+
 	@Autowired
 	private LoanService loanService;
-	
+	@Autowired
+	private AccountService accountService;
+
 	@GetMapping("Loans")
 	public List<Loan> getAllLoans() {
-		return loanService.getAllLoans();		
+		return loanService.getAllLoans();
 	}
- 
-	@GetMapping(path="Loans/{loanId}", produces = "application/json")	
+
+	@GetMapping("Accounts/{accountId}/Loans")
+	public List<Loan> getAllLoansByAccountId(@PathVariable("accountId") String accountId) {
+		Optional<Account> account = accountService.getAccount(accountId);
+		List<Loan> loans = new ArrayList<>();
+		if (account.isPresent()) {
+			loanService.getAllLoansByAccountId(account.get()).forEach(loans::add);
+		}
+		return loans;
+
+	}
+
+	@GetMapping("Accounts/{accountId}/Loans/{loanId}")
 	public Optional<Loan> getLoan(@PathVariable("loanId") String loanId) {
 		return loanService.getLoan(loanId);
-		
+
 	}
-	
-	@PostMapping("Loans")
-	public void addLoan(@RequestBody Loan loan) {
+
+	@PostMapping("Accounts/{accountId}/Loans")
+	public ResponseEntity<String> addLoan(@RequestBody Loan loan, @PathVariable("accountId") String accountId) {
+		Optional<Account> account = accountService.getAccount(accountId);
+		if (account.isPresent()) {
+			loan.setAccount(account.get());
+		}
 		loanService.saveLoan(loan);
+		return new ResponseEntity<>("Created", HttpStatus.CREATED);
 	}
-	
-	@PutMapping("Loans/{loanId}")
-	public void updateLoan(@RequestBody Loan loan, @PathVariable("loanId") String loanId) {
-		System.out.println("Loan:Object"+loan);
-		loanService.updateLoan(loanId,loan);
+
+	@PutMapping("Accounts/{accountId}/Loans/{loanId}")
+	public ResponseEntity<String> updateLoan(@RequestBody Loan loan, @PathVariable("loanId") String loanId,
+			@PathVariable("accountId") String accountId) {
+		Optional<Account> account = accountService.getAccount(accountId);
+		if (account.isPresent()) {
+			loan.setAccount(account.get());
+			loanService.saveLoan(loan);
+			return new ResponseEntity<>("Accepted", HttpStatus.ACCEPTED);
+		}
+		
+		return new ResponseEntity<>("Loan Not Found for Id"
+				+ ""
+				+ "::"+loanId+
+				" ", HttpStatus.NOT_FOUND);
 	}
-	
-	@DeleteMapping("Loans/{loanId}")
-	public void deleteLoan(@PathVariable("loanId") String loanId) {
-		loanService.deleteLoan(loanId);
+
+	@DeleteMapping("Accounts/{accountId}/Loans/{loanId}")
+	public ResponseEntity<String> deleteLoan(@PathVariable("loanId") String loanId) {
+		Optional<Loan> loan = loanService.getLoan(loanId);
+		if (loan.isPresent()) {
+			loanService.deleteLoan(loanId);
+			return new ResponseEntity<>("Deleted", HttpStatus.NO_CONTENT);
+		} else {
+			return new ResponseEntity<>("Loan Not Found for Id"
+					+ ""
+					+ "::"+loanId+
+					" ", HttpStatus.NOT_FOUND);
+		}
+
 	}
 }
